pyproject.toml
[tool.poetry]
name = "teste_palin"
version = "0.1.0"
description = ""
authors = ["Renan de Souza Carvalho <renan2.scarvalho@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8,<3.10"
fastapi = {extras = ["all"], version = "^0.70.0"}
pydantic = {extras = ["email"], version = "^1.8.2"}
dynaconf = "^3.1.7"
loguru = "^0.5.3"
bcrypt = "^3.2.0"
pre-commit = "^2.15.0"
motor = "^2.5.1"
pymongo = "^3.12.1"
geocoder = "^1.38.1"
sh = "^1.14.2"
sse-starlette = "^0.10.0"
termcolor = "^1.1.0"
selenium = "^4.1.0"
halo = "^0.0.31"
python-telegram-bot = "^13.10"
pika = "^1.2.0"
mkdocs = "^1.2.3"
mkdocs-material = "^8.2.5"
dateutils = "^0.6.12"
celery = {extras = ["redis"], version = "^5.2.7"}
flower = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
